requirements_full.txt
absl-py==2.4.0
accelerate==1.12.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
braceexpand==0.1.7
brotli==1.2.0
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
docstring_parser==0.17.0
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.46
gradio==6.4.0
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lmdb==1.7.5
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.90
orjson==3.11.5
packaging==26.0
pandas==2.3.3
peft==0.18.1
pillow==12.1.0
platformdirs==4.5.1
protobuf==6.33.5
psutil==7.2.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.1
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.17.0
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.52.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
starlette==0.50.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.24
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==5.0.0
triton==3.6.0
typed-argument-parser==1.11.0
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wandb==0.24.2
wcwidth==0.5.0
webdataset==1.0.2
Werkzeug==3.1.5
xformers==0.0.34
